requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
requests==2.28.1
speedtest-cli==2.1.3
urllib3==1.26.12
